+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix exception when delete returns no content-type
 * Add support for ruby 2.7.0
 * Add up/downgrading mailbox offer
+* Add dry run option to creating a mailbox
 
 ## Version 0.3.0